+++ src/app/src/main/java/ca/cmput301f21t22/nabu/model/LiveHabit.java
@@ -11,8 +11,6 @@ import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 
-import ca.cmput301f21t22.nabu.data.Occurrence;
-
 public class LiveHabit extends LiveDocument<Habit.Properties> implements Habit {
     @Nullable
     private Boolean shared;
