+++ src/app/src/main/java/ca/cmput301f21t22/nabu/model/LocalHabit.java
@@ -7,8 +7,6 @@ import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 
-import ca.cmput301f21t22.nabu.data.Occurrence;
-
 public class LocalHabit extends BaseObservable<Habit.Properties> implements Habit {
     @Nullable
     private Boolean shared;
